fix: Move update function to lib to avoid Next.js build issues
